03-multimodal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aidd"
version = "0.1.0"
description = "Telegram-бот с LLM (MVP)"
readme = "ReadMe.md"
requires-python = "==3.11.*"
dependencies = [
  "aiogram>=3.4,<4",
  "openai>=1.40,<2",
  "pydantic>=2,<3",
  "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aidd"]
